+++ src/backend/base/langflow/serialization/constants.py
@@ -1,0 +1,2 @@
+MAX_TEXT_LENGTH = 20000
+MAX_ITEMS_LENGTH = 1000
