+++ src/backend/base/langflow/utils/constants.py
@@ -173,6 +173,3 @@ MESSAGE_SENDER_AI = "Machine"
 MESSAGE_SENDER_USER = "User"
 MESSAGE_SENDER_NAME_AI = "AI"
 MESSAGE_SENDER_NAME_USER = "User"
-
-MAX_TEXT_LENGTH = 20000
-MAX_ITEMS_LENGTH = 1000
