+++ src/backend/tests/unit/utils/test_truncate_long_strings.py
@@ -1,7 +1,7 @@
 import math
 
 import pytest
-from langflow.utils.constants import MAX_TEXT_LENGTH
+from langflow.serialization.constants import MAX_TEXT_LENGTH
 from langflow.utils.util_strings import truncate_long_strings
