+++ src/backend/tests/unit/utils/test_truncate_long_strings_on_objects.py
@@ -1,5 +1,5 @@
 import pytest
-from langflow.utils.constants import MAX_TEXT_LENGTH
+from langflow.serialization.constants import MAX_TEXT_LENGTH
 from langflow.utils.util_strings import truncate_long_strings
